+++ src/config/interfaces/currency-config.interface.ts
@@ -1,0 +1,7 @@
+import { ExchangeRates } from 'mobx/model/system-config/exchange-rates';
+
+export interface CurrencyConfig {
+	getExchangeRate: (rates: ExchangeRates) => number;
+	decimals: number;
+	prefix: string;
+}